Add alert that stops access to unauthorized sites
The code for the implemented solution is commented out because it causes
conflicts with external links being loaded on the current website, such
as adds. A better solution needs to be found, but, for now, the written
code is within the scope of what has been asked for the challenge and
has been taught so far.